Moved Objects from entities.objects to core.objects Removed Debug element from TimeAlarmEvent Added methods with a forceReload to ResourceManager Optimized resource load on ResourceManager

+++ forscene-core/src/forscene/system/ISceneObject.java
@@ -97,7 +97,7 @@ public interface ISceneObject<T extends Layer> extends LayerObject<T>,
    *          the y
    * @return true, if successful
    */
-  abstract boolean contains(int x, int y);
+  abstract boolean contains(float x, float y);
 
   /**
    * Sets the update rate.
